requirements.txt
Pillow==8.4.0
stegano==0.9.9